fix(eco-14): cast u8→u64 for serde_json::Number and governor sampled_at_millis type

+++ crates/workflowd/src/governor.rs
@@ -445,7 +445,7 @@ impl Governor {
         };
 
         GovernorDecision {
-            sampled_at_millis: elapsed_micros / 1000,
+            sampled_at_millis: (elapsed_micros / 1000) as u64,
             cpu: CpuSnapshot {
                 available: sample.cpu.available,
                 quota_cores,
